requirements.txt
chainlit==1.1.404
pydantic==2.8.2
typing-extensions>=4.8.0
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2
azure-search-documents==11.4.0b11
azure-core==1.30.2
azure-storage-blob==12.23.1
pypdf==5.0.1
requests==2.32.3
reportlab==4.2.5
Pillow==10.4.0
pytesseract==0.3.10
langgraph
pdfminer.six==20231228
python-docx
